Blast
Started on fixing the blast problem in alignment. The blast.pl can be
put together with split.pl on a later point. Some of the code in
blast.pl is a bit strange becasue my brain took a break, will fix it
soon. And the psiblast.pl is a downloaded file that is used to do the
blast.