Can now find the gene
I tried to comment a lot. I haven't tried to realign because I can't run
that part. Not sure what parameters that should be used for it either.
First part find the gene and makes a sequence, the second part should
realign the "slice-part" and put the new  aligned sequence into the
alignment.
Also put the Acidianus genome file in the first map so we don't have to
worry about paths when testing